# docs/swagger.yaml
basePath: /
definitions:
...
  service.TransactionEntryForm:
    properties:
      account:
        type: string
      currency:
        type: string
      isAnotherCurrency:
        type: boolean
      number:
        type: number
      price:
        type: number
      priceCurrency:
        type: string
    required:
    - account
    type: object
  service.TransactionForm:
    properties:
      date:
        type: string
      desc:
        type: string
      divideDateList:
        items:
          type: string
        type: array
      entries:
        items:
          $ref: '#/definitions/service.TransactionEntryForm'
        type: array
      id:
        type: string
      narration:
        type: string
      payee:
        type: string
      rawText:
        type: string
      tags:
        items:
          type: string
        type: array
    required:
    - date
    - desc
    type: object
  service.TransactionTemplate:
    properties:
      date:
        type: string
      desc:
        type: string
      entries:
        items:
          $ref: '#/definitions/service.TransactionEntryForm'
        type: array
      id:
        type: string
      payee:
        type: string
      templateName:
        type: string
    required:
    - date
    - desc
    - templateName
    type: object
...
  service.UpdateRawTextTransactionForm:
    properties:
      id:
        type: string
      rawText:
        type: string
    required:
    - id
    - rawText
    type: object
...
host: localhost:10000
info:
  contact:
    name: BaoXuebin
  description: Beancount-GS 是一个基于 Beancount 的个人财务管理系统，提供账本管理、交易记录、统计分析等功能。
  title: Beancount-GS API
  version: "1.0"
paths:
...
  /api/auth/transaction:
    delete:
      description: 根据交易 ID 删除交易记录
      parameters:
      - description: 交易 ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 删除交易
      tags:
      - 交易
    get:
      description: 查询交易列表，支持按时间和账户筛选
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取交易列表
      tags:
      - 交易
    post:
      consumes:
      - application/json
      description: 新增交易，或传入 ID 更新已有交易
      parameters:
      - description: 交易表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.TransactionForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 新增或更新交易
      tags:
      - 交易
  /api/auth/transaction/batch:
    post:
      consumes:
      - application/json
      description: 批量写入多条交易记录到账本
      parameters:
      - description: 批量交易表单
        in: body
        name: body
        required: true
        schema:
          items:
            $ref: '#/definitions/service.TransactionForm'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 批量新增交易
      tags:
      - 交易
  /api/auth/transaction/detail:
    get:
      description: 根据交易 ID 获取交易详情（含分录）
      parameters:
      - description: 交易 ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取交易详情
      tags:
      - 交易
  /api/auth/transaction/payee:
    get:
      description: 返回最近交易中的商家名称列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取交易商家列表
      tags:
      - 交易
  /api/auth/transaction/raw:
    get:
      description: 根据交易 ID 获取原始文本
      parameters:
      - description: 交易 ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取交易原始文本
      tags:
      - 交易
    post:
      consumes:
      - application/json
      description: 根据交易 ID 更新原始文本
      parameters:
      - description: 更新原始文本表单
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateRawTextTransactionForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 更新交易原始文本
      tags:
      - 交易
  /api/auth/transaction/template:
    delete:
      description: 根据模板 ID 删除交易模板
      parameters:
      - description: 模板 ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 删除交易模板
      tags:
      - 交易模板
    get:
      description: 返回账本中保存的交易模板
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 获取交易模板列表
      tags:
      - 交易模板
    post:
      consumes:
      - application/json
      description: 新增一条交易模板配置
      parameters:
      - description: 交易模板
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.TransactionTemplate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - LedgerId: []
      summary: 新增交易模板
      tags:
      - 交易模板
...
securityDefinitions:
  LedgerId:
    description: 需要授权的接口请在 Header 中传入 ledgerId
    in: header
    name: ledgerId
    type: apiKey
swagger: "2.0"